requirements.txt
###### Requirements without Version Specifiers ######
twitchAPI
python-dotenv